+++ src/main/java/com/jonathan/futebol_api/adapter/dto/EstadioResponseDTO.java
@@ -1,6 +1,6 @@
 package com.jonathan.futebol_api.adapter.dto;
 
 public record EstadioResponseDTO(
-        Integer idEstadio,
+        Long idEstadio,
         String nome
 ) {}
